Made homepage the main page
User's directly go to homepage instead of login
